backend/src/inventory/views.py
from rest_framework import viewsets

from inventory.models import InventoryItem
from inventory.serializers import InventoryItemSerializer


class InventoryItemViewSet(viewsets.ModelViewSet):
    queryset = InventoryItem.objects.all()
    serializer_class = InventoryItemSerializer
