backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Inventory backend service built with Django and uv."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.2.11",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.5",
]
test = [
    "pytest>=8.3.3",
    "pytest-django>=4.11.1",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "inventory_project.settings"
pythonpath = ["src"]
